# docs/swagger.yaml
basePath: /
definitions:
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
  models.ShowUserData:
    properties:
      id:
        example: 1023456789
        type: integer
      name:
        example: NKTKLN
        type: string
      username:
        example: nktkln
        type: string
    type: object
...
info:
  contact:
    email: nktkln@nktkln.com
    name: API Support
    url: https://nktkln.com
  description: some description
  license:
    name: MIT
    url: http://github.com/
  title: ToDo API
  version: "1.0"
paths:
...
  /user/show/data-by-id:
    get:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowUserData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get basic user data
      tags:
      - Show user data
  /user/show/data-by-token:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowUserData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Get basic user data
      tags:
      - Show user data
  /user/show/icon:
    get:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Get user icon
      tags:
      - Show user data
securityDefinitions:
  token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"